+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+Breaking changes:
+
+- Function implementations (class constructors) must conform to the `ImplementableFunctionImplementationConstructor` type rather than `ClassConstructor` from the `class-transformer` package.
+
+Features:
+
+- Provide the `createContext` and `registerMockFunction` testing utilities.
+
 ## v0.2.1 (2023-05-05)
 
 Fixes: